refactor: split storage and chart logic into modules
