+++ index.html
@@ -320,6 +320,6 @@
   </dialog>
 
   <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-  <script src="app/js/app.js"></script>
+  <script type="module" src="app/js/app.js"></script>
 </body>
 </html>
